requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
attrs==23.1.0
Automat==22.10.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
decorator==5.1.1
ecdsa==0.18.0
executing==2.0.1
fastapi==0.104.1
filelock==3.13.1
greenlet==3.0.1
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipdb==0.13.13
ipython==8.17.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
lxml==4.9.3
matplotlib-inline==0.1.6
mysql-connector-python==8.2.0
outcome==1.3.0.post0
packaging==23.2
parsel==1.8.1
parso==0.8.3
passlib==1.7.4
prompt-toolkit==3.0.41
Protego==0.3.0
protobuf==4.21.12
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dotenv==1.0.0
python-jose==3.3.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
rsa==4.9
Scrapy==2.11.0
selenium==4.15.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
tldextract==5.1.0
traitlets==5.13.0
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
w3lib==2.1.2
wcwidth==0.2.10
wsproto==1.2.0
zope.interface==6.1
